requirements.txt
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorclass==2.2.2
compressed-rtf==1.0.6
cryptography==36.0.1
cycler==0.11.0
docopt==0.6.2
easygui==0.98.2
ebcdic==1.1.1
extract-msg==0.30.8
Flask==2.0.2
fonttools==4.29.0
idna==3.3
imageio==2.14.1
IMAPClient==2.2.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
lark-parser==0.12.0
lmdb==1.3.0
MarkupSafe==2.0.1
matplotlib==3.5.1
msoffcrypto-tool==5.0.0
natsort==8.0.2
networkx==2.6.3
numpy==1.22.1
olefile==0.46
oletools==0.60
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.4.0
pcodedmp==1.2.6
pdf2image==1.16.0
pdfkit==1.0.0
Pillow==9.0.0
pipreqs==0.4.11
pycparser==2.21
pyparsing==2.4.7
PyPDF2==1.26.0
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==6.0
requests==2.27.1
RTFDE==0.0.2
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
thefuzz==0.19.0
threadpoolctl==3.0.0
tifffile==2021.11.2
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
Werkzeug==2.0.2
yarg==0.1.9
statistics==1.0.3.5